Add null checking on EventManager::dispatchEvent
Ftc::Core::EventManager::dispatchEvent does not check if the Event *event is allocated properly, and may result in null reference. Since the Event::create uses new (std::nothrow) on allocation of class Event, the return value of Event::create may be nullptr on case of bad allocation, thus should be checked on caller-side.